Separate directory for analyzers
