+++ analyzer/codechecker_analyzer/analyzers/clangsa/config_handler.py
@@ -15,7 +15,7 @@ import re
 
 from codechecker_common.logger import get_logger
 
-from . import config_handler
+from .. import config_handler
 
 LOG = get_logger('analyzer.clangsa')
