+++ analyzer/codechecker_analyzer/analyzers/clangsa/ctu_triple_arch.py
@@ -12,9 +12,10 @@ from __future__ import absolute_import
 
 import shlex
 
-from ..analyzer_env import extend_analyzer_cmd_with_resource_dir
+from codechecker_analyzer.analyzer_env import \
+    extend_analyzer_cmd_with_resource_dir
 
-from . import analyzer_base
+from .. import analyzer_base
 
 
 def get_compile_command(action, config, source='', output=''):
