+++ analyzer/codechecker_analyzer/analyzers/clangtidy/__init__.py
@@ -1,0 +1,5 @@
+# -------------------------------------------------------------------------
+#                     The CodeChecker Infrastructure
+#   This file is distributed under the University of Illinois Open Source
+#   License. See LICENSE.TXT for details.
+# -------------------------------------------------------------------------
