+++ analyzer/codechecker_analyzer/analyzers/clangtidy/output_converter.py
