+++ analyzer/tests/unit/test_tidy_output_converter.py
@@ -21,7 +21,8 @@ try:
 except ImportError:
     from io import BytesIO as StringIO
 
-import codechecker_analyzer.tidy_output_converter as tidy_out_conv
+import codechecker_analyzer.analyzers.clangtidy.output_converter as \
+    tidy_out_conv
 
 OLD_PWD = None
